build(core): replace -fsingle-precision-constant with explicit casts
The option -fsingle-precision-constant is unsupported in clang. Instead,
I'm adding -Wdouble-promotion (no hits in the codebase) and
-Wfloat-conversion (fixing the warning sites).

-fsingle-precision-constant ensures that all double literals without 'f'
suffix are treated as floats. But we don't care about literals per se,
we care about silent double promotions for calculations. The new set of
warnings is portable and more closely matches the intent.

+++ core/embed/io/display/unix/display_driver.c
@@ -681,7 +681,7 @@ static void display_draw_suspend_overlay_internal(void) {
   // create a blue texture
   SDL_Texture *overlay =
       SDL_CreateTexture(drv->renderer, SDL_PIXELFORMAT_RGBA8888,
-                        SDL_TEXTUREACCESS_STATIC, screen.w, screen.h);
+                        SDL_TEXTUREACCESS_STATIC, (int)screen.w, (int)screen.h);
   SDL_SetTextureBlendMode(overlay, SDL_BLENDMODE_BLEND);
   SDL_SetRenderTarget(drv->renderer, overlay);
